Integrate UserExtendedClipboardAction into UserClipboardAction
