+++ com.woltlab.wcf/clipboardAction.xml
@@ -26,7 +26,7 @@
 		</action>
 		
 		<action name="merge">
-			<actionclassname><![CDATA[wcf\system\clipboard\action\UserExtendedClipboardAction]]></actionclassname>
+			<actionclassname><![CDATA[wcf\system\clipboard\action\UserClipboardAction]]></actionclassname>
 			<showorder>4</showorder>
 			<pages>
 				<page><![CDATA[wcf\acp\page\UserListPage]]></page>
@@ -34,7 +34,7 @@
 		</action>
 		
 		<action name="enable">
-			<actionclassname><![CDATA[wcf\system\clipboard\action\UserExtendedClipboardAction]]></actionclassname>
+			<actionclassname><![CDATA[wcf\system\clipboard\action\UserClipboardAction]]></actionclassname>
 			<showorder>5</showorder>
 			<pages>
 				<page><![CDATA[wcf\acp\page\UserListPage]]></page>
